+++ packages/parchment/test/codecs/html_test.dart
@@ -1,5 +1,5 @@
 import 'package:parchment/parchment.dart';
-import 'package:parchment/src/convert/html.dart';
+import 'package:parchment/src/codecs/html.dart';
 import 'package:quill_delta/quill_delta.dart';
 import 'package:test/test.dart';
